chore(use-show): use `useResourceParams` hook
